requirements.txt
alembic==1.6.2
blinker==1.4
click==8.0.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
idna==3.1
importlib-metadata==4.0.1
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
pkg-resources==0.0.0
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.15
typing-extensions==3.10.0.0
visitor==0.1.3
Werkzeug==2.0.0
WTForms==2.3.3
zipp==3.4.1
